requirements.txt
pandas==2.1.4
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
schedule==1.2.0
loguru==0.7.2
pydantic==2.5.2
openpyxl==3.1.2

